+++ src/main/java/com/macd/ps/model/PaymentDto.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @NoArgsConstructor
@@ -15,5 +16,5 @@ public class PaymentDto {
     private long id;
     private long orderNumber;
     private LocalDateTime paymentDate;
-    private double totalAmount;
+    private BigDecimal totalAmount;
 }
